add two profiles and some quick fixes (security rules, exception handlers, refactoring)

+++ src/main/java/com/griddynamics/shopapi/dto/OrderItemDto.java
@@ -19,6 +19,6 @@ public class OrderItemDto {
   public OrderItemDto(OrderItem item) {
     productId = item.getProductId();
     quantity = item.getQuantity();
-    price = item.getPrice();
+    price = item.getPrice().doubleValue();
   }
 }
